moving mapper function from domain & proper directory for favorite domain package
