+++ core/src/main/java/com/edwinyosua/core/data/repository/GameFavoriteRepository.kt
@@ -1,8 +1,8 @@
 package com.edwinyosua.core.data.repository
 
 import com.edwinyosua.core.data.local.LocalDataSources
-import com.edwinyosua.core.domain.favorite.GameListFavorite
-import com.edwinyosua.core.domain.favorite.IGameFavoriteRepository
+import com.edwinyosua.core.domain.favorite.model.GameListFavorite
+import com.edwinyosua.core.domain.favorite.repository.IGameFavoriteRepository
 import com.edwinyosua.core.utils.DataMapper
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
