+++ core/src/main/java/com/edwinyosua/core/data/repository/GameListRepository.kt
@@ -2,9 +2,9 @@ package com.edwinyosua.core.data.repository
 
 import com.edwinyosua.core.data.remote.network.ApiResponse
 import com.edwinyosua.core.data.remote.network.ApiService
-import com.edwinyosua.core.domain.home.mapper.toDomain
 import com.edwinyosua.core.domain.home.model.GamesList
 import com.edwinyosua.core.domain.home.repository.IGameListRepository
+import com.edwinyosua.core.utils.toDomain
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
