+++ core/src/main/java/com/edwinyosua/core/di/CoreModule.kt
@@ -8,7 +8,7 @@ import com.edwinyosua.core.data.repository.GameDetailRepository
 import com.edwinyosua.core.data.repository.GameFavoriteRepository
 import com.edwinyosua.core.data.repository.GameListRepository
 import com.edwinyosua.core.domain.detail.repository.IGameDetailRepository
-import com.edwinyosua.core.domain.favorite.IGameFavoriteRepository
+import com.edwinyosua.core.domain.favorite.repository.IGameFavoriteRepository
 import com.edwinyosua.core.domain.home.repository.IGameListRepository
 import com.edwinyosua.core.ui.GameFavListAdapter
 import com.edwinyosua.core.ui.GameListAdapter
