--- core/src/main/java/com/edwinyosua/core/domain/detail/mapper/GameDetailMapper.kt
@@ -1,9 +1,0 @@
-package com.edwinyosua.core.domain.detail.mapper
-
-import com.edwinyosua.core.data.remote.response.GameDetailResponse
-import com.edwinyosua.core.domain.detail.model.GameDescription
-
-fun GameDetailResponse.toDomain(): GameDescription =
-    GameDescription(
-        description = this.descriptionRaw
-    )
