+++ core/src/main/java/com/edwinyosua/core/domain/favorite/repository/IGameFavoriteRepository.kt
@@ -1,5 +1,6 @@
-package com.edwinyosua.core.domain.favorite
+package com.edwinyosua.core.domain.favorite.repository
 
+import com.edwinyosua.core.domain.favorite.model.GameListFavorite
 import kotlinx.coroutines.flow.Flow
 
 interface IGameFavoriteRepository {
