+++ core/src/main/java/com/edwinyosua/core/domain/favorite/usecase/GameFavoriteUseCase.kt
@@ -1,5 +1,6 @@
-package com.edwinyosua.core.domain.favorite
+package com.edwinyosua.core.domain.favorite.usecase
 
+import com.edwinyosua.core.domain.favorite.model.GameListFavorite
 import kotlinx.coroutines.flow.Flow
 
 interface GameFavoriteUseCase {
