+++ favorite/src/main/java/com/edwinyosua/favorite/FavoriteViewModel.kt
@@ -2,7 +2,7 @@ package com.edwinyosua.favorite
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.asLiveData
-import com.edwinyosua.core.domain.favorite.GameFavoriteUseCase
+import com.edwinyosua.core.domain.favorite.usecase.GameFavoriteUseCase
 
 class FavoriteViewModel(gameFavRepo: GameFavoriteUseCase) : ViewModel() {
